[Buffers] No constraints for mem.-iface.-adjacent channels
This fixes a bug in the FPL'22 buffer placement algorithm that could
yield unsatisfiable MILPs when placing buffers on channels outside all
CFDFCs. Channels adjacent to memory interfaces (MCs, LSQs) must be
completely ignored both in channel and unit constraints to ensure that
the MILP is satisfiable (these channels are anyway constrained to be
unbufferizable by our constraint-setting pass). This was an oversight
since our FPGA'20 buffer placement algorithm follows this exact same
logic, for the same reason.