[Speculation] Pass skeleton with JSON-based unit placement (#60)
This commit introduces an initial version of the experimental
`HandshakeSpeculation` pass, which is the first step in supporting
speculative execution in dataflow circiots. The pass places
speculation-related operations on specific channels in the IR which are
encoded in an input JSON file, whose format mirrors the operand
attribute annotation format from the back-annotation pass. The
`SpeculationPlacements` helper class holds the placement information
and is initialized by the JSON-parsing logic.

This first speculation-related commit includes the basic pass structure
and some dummy connections, and will be followed by further
contributions implementing more features. The idea is the following.

- **Commit 1 (this one).**: the pass will place the units according to
an input JSON file, and use simple connections.
- **Commit 2**: use the correct control path for connecting Commits and
SaveCommits units. Still read position from JSON.
- **Commit 3**: if no JSON is provided, automatically choose the
positions for placing the units.

The work for the two next commits is mostly done and being tested.

The pass can be ran through the optimized using the following syntax.  
`--handshake-speculation="json-path=positions.json"`

The input JSON file should have the following format, denoting channels
in the IR on which to place certain speculation-related operations.
```json
{
  "speculator": {
    "operation-name": "fork5",
    "operand-idx": 0
  },
  "saves": [
    {
      "operation-name": "mc_load0",
      "operand-idx": 0
    },
    {
      "operation-name": "mc_load1",
      "operand-idx": 0
    }
  ],
  "commits": [
    {
      "operation-name": "cond_br0",
      "operand-idx": 1
    }
  ],
  "save-commits": [
    {
      "operation-name": "buffer10",
      "operand-idx": 0
    }
  ]
}
```

Example JSON files will be explicitly added in a next commit to drive
unit tests for the pass.