[integration-test] Move header to Dynamatic's include folder
This moves the header included by all integration tests (previously,
`integration_utils.h`) from the `integration-test` folder to the
`include/dynamatic` folder (the standard location for all our headers)
and renames it to `Integration.h`. This is done for consistency and to
avoid having an explicitely relative include path in our integration
tests' sources. The latters are modified to reflect the change.

The simulation script used by the Dynamatic frontend is updated to
reflect this change (the HLS verifier needs to see the header to compile
and test the kernel).