[dynamatic] Add standard include directory to profiling build
This commit adds the standard Dynamatic include directory to the list of
include directories looked up for building the kernel in "profiling
mode" during compilation. Without this, the build script fails to find
the `dynamatic/Integration.h` header included by all of our integration
tests. This change should have happened in the previous commit but
was not due to an oversight.