[FuncMaximizeSSA] SSA maximization pass
This moves the `MaximizeSSA` pass from CIRCT to Dynamatic with cosmetic
modifications that do not impact its fundamental behavior but make it
slightly easier to use (which we will soon do from the `cf-to-handshake`
conversion pass).

The pass puts the body of a func-level function into maximal SSA form,
ensuring that all SSA values are used in the block where they are
defined (either through a block argument or the result of an operation
within the block).