[StandardToHandshakeFPGA18] Use Dynamatic's SSA maximization
This makes the `StandardToHanshakeFPGA18` pass use the newly introduced
Dynamatic version of the SSA maximization pass, which is slightly more
convenient to use than its CIRCT counterpart. This in turns allows us to
rely less on the CIRCT's version of the conversion pass, by removing the
dependency on the `partiallyLowerRegion` method.