[export-vhdl] Integrate legacy `dot2vhdl` in Dynamatic
This commit integrates the legacy `dot2vhdl` tool into Dynamatic,
cleaning it up a fair bit and removing things we do not use in the
process. The frontend `write-hdl` command now uses this new integrated
version instead of the old one.

This is only meant as a temporary measure so that Dynamatic users may
use the tool witout having to install legacy Dynamatic on the side. The
tool is renamed `export-vhdl` to be consistent with similar tools
already present on the repository.