[export-vhdl] Move LSQ generation out of tool
This commit removes the LSQ generation step from the `export-vhdl`
binary and moves it to the `write-hdl` frontend script. This is most
likely a temporary measure while we decide where the LSQ generator will
be located (on legacy Dynamatic's repository or on this repository)
and/or which one we will use (old Chisel one, new Python one).