[hls-verifier] Integrate legacy `hls_verifier` in Dynamatic
This commit integrates the legacy `hls_verifier` tool into Dynamatic,
cleaning it up a fair bit and removing things we do not use in the
process. The frontend `simulate` command now uses this new integrated
version instead of the old one.

The tool is renamed `hls-verifier` to be consistent with similar tools
already present on the repository.