[export-vhdl] Fix string to integer converion behavior
This is a simple fix to the string to integer conversion logic used in
the `export-vhdl` tool (legacy `dot2vhdl` tool). Some trickery needs to
happen on our side to emulate legacy's exception-based behavior
successfully without exceptions. The resulting code is not something to
be proud of, yet makes our benchmarks work.