[hls-verifier] Use C++ macro for IO generation
This commit moves the input vectors/golden outputs generation logic out
of the `hls-verifier` and in the `CALL_KERNEL` C macro already used in
all of our integration tests. Similarly to the profiling inputs
ultimately used by our smart buffer placement algorithm, a boolean macro
is added to the `Integration.h` header to control what happens before
and after the actual kernel call. In this case, we add logic to log the
content of all kernel arguments before and after the call (as well as
the kernel's return value, if there is one) to files on disk, according
to the structure expected by the hls-verifier for co-simulation.

This change adds a constraint to the way the `CALL_KERNEL` macro is
called: function arguments have to be named exactly the same as the
parameter names in the function's declaration, since they are used
internally by the macro to derive the filename in which each argument's
content is stored. This is imposed by the `hls-verifier`'s internal
logic, and may change in the future to be more flexible (for example,
by basing the filename on the order of arguments instead of their names,
since this necessarily matches the function's declaration).

This has the nice side-effect of removing the tool's unstated dependency
on gcc, since we know use the version of clang we build to compile the
instrumented benchmarks and extract input/output values.