[CIRCT] Remove useless references to the `circt` namespace
This commit begins the process of removing our CIRCT dependency by
getting rid of easily removable references to the `circt` namespace.
Ultimately, we will transfer the CIRCT infrastructure we still rely on
to the `dynamatic` namespace, so references to entities in the `circt`
namespace will later point to the `dynamatic` namespace.