[StandardToHandshakeFPGA18] Stop relying on CIRCT lowering steps
This commit moves all the code directly related to the
`StandardToHandshake` conversion pass into Dynamatic itself, instead of
having some lowering steps live in CIRCT while the rest of the pass
lives in Dynamatic. This involves some code copying from CIRCT to
Dynamatic and some cosmetic improvements to the pass's code.

This continues to prepare the transition away from CIRCT.