[HW] Move dialect definition in Dynamatic
This commit "moves" the HW dialect from CIRCT to Dynamatic. All files
related to the dialect's definition and implementation are copied
directly inside the Dynamatic repository, and all references to the
dialect throughout the codebase are changed so that the "Dynamatic
version" of the dialect is used intead of the "CIRCT version". At the
same time, we remove parts of the dialect (e.g., operations, interfaces)
that we do not use in Dynamatic. More cleanup and refactors will likely
follow. This is the last significant step towards moving away from CIRCT
entirely.