Farewell CIRCT
This removes CIRCT as a submodule to Dynamatic. It served us very well,
and we will continue to benefit greatly from the Handshake and HW
dialects. However, a divergent evolution of Handshake and the
progressive removal of CIRCT passes from our compilation flow makes the
burden of bundling two different versions of LLVM/MLIR with the project
(one for Polygeist and one for CIRCT) too great with respect to the
benefits of receiving CIRCT upstream updates.

While practical concerns force us to part ways for now, I do hope that
parts of Dynamatic and CIRCT can somehow converge back at some point, as
I believe it would be very beneficial to the EDA open-source community.

Dynamatic now builds against the version of LLVM/MLIR that is a
submodule to Polygeist. To increase flexibility, the Polygeist submodule
now tracks a fork of upstream Polygeist instead of the official
repository. This will allow us to occasionally modify Polygeist itself
or bump the LLVM version it/Dynamatic builds against.

The build infrastructure, VSCode config files, and top-level README are
modified to reflect these changes. At least we build faster and smaller!