[CfToHandshake] Fix branch insertion issues
After the transition away from CIRCT and the change in underlying LLVM
version we started to have issues related to the branch insertion logic
in the `lower-cf-to-handshake` conversion pass. In particular, live-outs
to a block that would be used multiple times (either by a single
successor block or by multiple successor blocks) would not be wired
correctly with the rest of the circuit, leading to useless conditional
branches and/or incorrect data operands for merge-like operations.

According to my understanding, these should have appeared earlier,
though they only manifested themselves now. This may be due to internal
changes in LLVM. In any case, the code is now "more correct" than it was before.