[CfToHandshake] Fix inconsistent branch insertion order
The last change in the `lower-cf-to-handshake` pass introduced
non-determinism in the produced IR due to the use of a data-structure
(`DenseSet`) that does not guarantee a deterministic iteration order in
the branch insertion logic. This commit replaces it with a semantically
identical data-structure (`SetVector`) that guarantees iteration order,
yielding a deterministic output IR for any input.

Unit tests for the pass are updated to reflect recent changes. The pass
no longer generates useless conditional branches, making resulting IRs
shorter in some situations.