[Support] Improve edge-styling in `DOTPrinter`
This adds a couple cosmetic attributes to the DOT edges output by our
DOT printer, making the arrows' endpoints different based on the type of
connection they represent. In particular, this changes the styling of
edges representing the condition operand of conditional branches and the
select operand of muxes to make it easier to distinguish these "unique"
input channels from others in the DOT visualization.