[Handshake] `DynamaticReturnOp` -> `ReturnOp`
Renames the `handshake::DynamaticReturnOp` operation to
`handshake::ReturnOp`. Its mnemonic is also changed from `d_return` to
`return`. The decision to originally prepend `Dynamatic` to the
operation's name was to disambiguate with another CIRCT operation in
Handshake named `ReturnOp`. Now that the Handshake dialect exists
directly as part of Dynamatic and that we have deleted the CIRCT
return operation, we can rename ours for brevity.