[visual-dataflow] Add wlf2csv Python script
The `wlf2csv.py` script transforms the Modelsim waveform output after
simulation of a dataflow circuit into a human-readable and
machine-parsable CSV-formatted sequence of channel state changes that
are read by the dataflow visualizer to create an interactive
visualization of the simulated circuit. It takes as input the DOT file
that was originally used to generate the simulated VHDL as well as
Modelsim's waveform output (`*.wlf`), and produces the CSV in a provided
output folder.