[dynamatic] Add `visualize` command
This adds the `visualize` command (along with an accompanying script) to
the list of commands understood by the frontend. The command creates
input files required by the dataflow visualizer: a DOT file with added
positioning information (from the original DOT file) and a CSV file with
the list of dataflow channel state changes (from the Modelsim-generated
waveform file).