[dynamatic] Remove legacy dependency for `write-hdl`
The `write-hdl` command of Dynamatic's frontend only depended on legacy
for the LSQ generator, a 50MB .jar file stored on legacy's repository.
This commit removes this dependency by instead downloading the LSQ
generator from legacy's GitHub repository using curl the first time it
is needed. Subsequent calls to `write-hdl` are then able to use the
generator directly. The generator is downloaded in `data/vhdl/LSQ.jar`.