[tutorials] Introduce UsingDynamatic tutorial
This adds a new tutorial to the repository's documentation, meant to be
the first exercise of FPGA'24's Dynamatic workshop. The tutorial is
meant to guide new users through their first interaction with the
framework by having them synthesize and simulate a dataflow circuit from
C code using Dynamatic's frontend. At the end, users are invited to
visualize the resulting dataflow circuit using the dataflow visualizer.
This last part still needs some work, but the rest of the tutorial
should be in a pretty good state.