[tutorials] Introduce `ModifyingDynamatic` tutorial
This introduces the second part of our "Introduction to Dynamatic"
tutorial, which guide readers through the design of a small optimization
pass and its ultimate debugging using our dataflow visualizer. Some
edits are made to the first chapter of this tutorial to fix a couple
issues and make it more consistent with the second part.