[CMake] Fix CMake arguments in build script
This commit slightly modifies the CMake arguments given to
LLVM/Polygeist/Dynamatic to make more sense in each respective mode that
the project can be built in (Debug or Release). Assertions are now
removed in Release mode (they were not before), and LLVM/Polygeist will
use LLD as a linker by default (unless -o is provided).