[Buffers] Simple pass for inserting a single buffer (#62)
This commit introduces an experimental pass
`--handshake-placebuffers-custom` which inserts a single buffer in the
IR on a channel referenced by its producer's name and result index (as
such, the IR must be materialized as a pre-condition to the pass). 

This pass facilitates externally prototyping a custom buffer placement
analysis, e.g., in Python. This also makes the results of some research
artifacts (e.g.,
[Mapbuf](https://github.com/Nozidoali/MapBuf/tree/master)) developed in
Python easily reproducible in the current Dynamatic framework.

A typical workflow would be:

1. Export the `handshake_transformed.mlir` to DOT, apply some external
analysis using Python to determine where to place buffers.
2. Using this pass: add all the buffers, and generate
`handshake_buffered.mlir`.
3. Continue with the rest of the HLS flow.

For example, we can add buffers by as follows (this is for the benchmark
`fir`):

```sh
dynamatic-opt handshake_transformed.mlir \
	--handshake-placebuffers-custom="pred=mux1 outid=0 slots=1 type=oehb" \
	--handshake-placebuffers-custom="pred=mux1 outid=0 slots=1 type=tehb" \
	--handshake-placebuffers-custom="pred=mux2 outid=0 slots=1 type=oehb" \
	--handshake-placebuffers-custom="pred=mux2 outid=0 slots=1 type=tehb" \
	--handshake-placebuffers-custom="pred=control_merge2 outid=0 slots=1 type=oehb" \
	--handshake-placebuffers-custom="pred=control_merge2 outid=0 slots=1 type=tehb" \
	> "handshake_buffered.mlir"

dynamatic-opt handshake_buffered.mlir --handshake-canonicalize \
  > handshake_export.mlir
```

---------

Co-authored-by: Jiahui Xu <jxu@ethz.ch>