[VSCode] Clean-up C/C++/clangd configuration
This commit cleans up the VSCode configuration for Dynamatic, which
relies on Microsoft's basic C++ extension for basic Intellisense and on
LLVM's clangd extension for static analysis/formatting and more advanced
features.

In particular, clangd needs a file called `compile_commands.json` to be
present in the project's build folder to know how to compile each of our
source files for proper linting. This file is automatically generated by
CMake if it's given the `-DCMAKE_EXPORT_COMPILE_COMMANDS` flag, which
the project's build script now provides.