[dynamatic] Change script extension `.sh` -> `.dyn`
This changes the file extension used by all frontend scripts throughout
the repository from `.sh` to `.dyn` to make it clearer that these are
not shell scripts but sequences of commands readable by Dynamatic's
frontend. The confusion was especially clear in the recently introduced
tutorial, which has actual shell scripts located next to "Dynamatic
scripts".