[tutorials] Reflect recent changes in documentation
This mainly updates the documentation of the "Introduction" tutorial,
reflecting recent changes in file layout and file paths and fixing a
couple typos along the way. Frontend scripts for the tutorial are also
updated to reflect filepath changes. A .gitignore is also added to the
tutorial's subfolder so that git ignores output directories generated
during the tutorial.