[visual-dataflow] Identical debug/release library name
This changes the name of the C++ library/GDExtension so that it no
longer contains "debug" or "release" in it depending on the compile mode
it was created in. This makes management of this library simpler from
Godot's side (which, in editor, seems to always expect the "debug"
version of the library anyway), especially to export the final binary.
It is now up to the user to decide whether they want the library to be
compiled in Debug or Release mode, independently from whether they want
the *Godot project* itself to be compiled in Debug or Release mode. The
`.gdextension` file is modified to reflect the change.