[visual-dataflow] Directly open graph on launch
This commits adds the possibility of opening directly the dataflow
visualizer on a specific graph by providing paths to the DOT and CSV
files normally selected by the user using the GUI as command-line
arguments. This smoothes out the experience considerably when using
Dyamatic's frontend, which now directly opens the visualizer on the
circuit at hand with the `visualize` command.