[visual-dataflow] Make visualizer's CMake independent
This makes the `CMakeLists.txt` file in `visual-dataflow` independent
from Dynamatic's top-level CMake file, which means that a separate call
to `cmake` is necessary to build the dataflow visualizer after having
built Dynamatic. The dataflow visualizer is not part of the core
compiler and there are use cases where one would not wish to build it so
it makes more sense for it to be a "separate tool" in the build process.
The `build.sh` script accepts a `-v/--visual-dataflow` flag to enable
this separate build step, simialrly to before. Prior to the change
however, and since `visual-dataflow` was part of Dyanamtic's cmake,
changes to the tool's build state would require the user to
also provide the `-f/--force` flag to take the new state into account.
This is no longer necessary.

Note that the build script will, at this point, only build the
GDExtension that the Godot project uses, but will not export the Godot
project itself to a binary. This additional build step may be added in
the future.