[visual-dataflow] Node/Edge visual improvements
This makes several small yet noticeable improvements to the dataflow
visualizer. In particular, all lines in the visualizer are made 50%
thicker and have anti-aliasing enabled, making it look closer to what a
tool like `xdot` generated out-of-the-box. Dashed lines are also changed
to be more "dotted" (shorter line segments, wider spacing between
segments) and oval nodes are made smoother by increasing the number of
vertices used to draw them. Finally, the edge and node drawing code is
made slightly cleaner overall.

The frontend's `visualize` command is also improved so that, prior to
regenerating the DOT with positioning information, it makes sure to
change the DOT's edge style to "ortho" instead of "spline".