[visual-dataflow] Add Godot export to build script
Enables export of the Godot project corresponding to the visual-dataflow
tool from the top-level build script through the `--export-godot
<godot-path>` optional flag. As indicated, the command requires that the
engine is present on the user's machine at the indicated path (which may
be absolute or relative). The command creates the tool's binary at
`visual-dataflow/bin/visual-dataflow` and creates a symlink to it at
`bin/visual-dataflow`. It is still required that the user has opened the
project in the editor at least once to download the appropriate export
libraries. At this point, only Linux debug builds are supported.