[tutorials] Visualizer now triggers automatically
This updates the documentation of both chapters of the "Introduction"
tutorial to reflect the fact that the interactive dataflow visualizer
now opens automatically when the `visualize` command is issued to the
frontend. Users no longer have to launch the binary and input its
sources files manually, which improves the tutorial's overall flow.