[visual-dataflow] Improve edge color handling
This commit is an overall improvement over the handling of edge colors
in the interactive dataflow visualizer. The C++ and GDScript code that
handles the redrawing and recoloring of edges depending on the current
cycle and channel state is cleaned up and improved. Default colors for
each channel are also changed in an attempt to make them more
"meaningful" (e.g., "transfer" becomes green by default).

The Godot scene representing the legend sees some of its nodes renamed
(as well as their associated event-handling functions) and reorganized
for clarity. The edge colors in the graph are now updated live whenever
the user drags the pointer in any of the color pickers, as opposed to
updating the colors only when the picker is closed.