[tutorials] Change loop bound from 10 to 8 in "Introduction"
This changes the loop's upper bound from 10 to 8 in both kernels of the
"Introduction" tutorial. This avoids an overflow in the `loop_multiply`
kernel which would otherwise return 0 after 5 squaring of the initial
value 2, which is confusing. The second kernel's loop bound did not
create any confusion but it is more consistent for it to also be 8 given
that it does not impact the exercise in any way.