[Speculation] Automatic speculative units placement (#66)
This commit enables the speculative execution pass
(`--handshake-speculation`) to automatically place speculative units in
a Handshake function without the user having to manually specifying
their positions in a JSON file.

The pass can now be called with
`--handshake-speculation="json-path=myjson.json automatic=true"` to
instruct the pass to automatically derive placements for speculative
units. Notice that `automatic` is `true` by default, so to manually
specify the units, ones must explicitly specify `automatic=false`.

The feature is implemented in the `PlacementFinder` class, whose
constructor and `findPlacements` method are publicly exposed. The job of
finding positions for all types of speculative units is given to private
methods of the class.

This commit also contains new unit tests to verify the feature's
behavior.

Part of the code included in this contribution will in the future become
part of our general-purpose CFG analysis infrastructure.