[docs] Add advanced build instructions
This adds a new Markdown document in the official Dynamatic
documentation which provides instructions for "advanced" users to build
the project. In particular, it goes over the main flags/arguments
offered by the top-level build script and explain the partially manual
process of building the interactive dataflow circuit visualizer along
the rest of Dynamatic.

The `GettingStarted.md` document now links to the advanced build
instructions in its subsection on building the project.