[README] Thin out build instructions and link to advanced ones
The build isntructions present in the top-level README file are thinned
out so that they are more straightforward to follow for users who may
only wish to use Dynamatic without modifying the code. In that spirit,
the README now invites users to build the project in Release mode
instead of Debug by default in an effor to minimize build size and time.

Users who wish to modify Dynamatic or personalize the build process to
their needs are now invited to follow the newly added "advanced build
instructions", which is linked from the README.