[Support] Move `BBArc` infra. to CFG support (#72)
In #66, logic for determining a BB's predecessor arcs was added as part
of the speculation pass. Here, the logic is moved into generic CFG
support to allow other parts of the codebase to use it. In addition,
some C++ STL data-structures were changed to semantically equivalent
LLVM/MLIR data structures.