[dynamatic] Improve frontend comment handling
This improves the way the frontend handle comments, both in interactive
and non-interactive mode. Every character on a line after a '#'
character is now completely ignored by the command processing logic.

This commit also fixes a "visual glitch" where lines starting with a '#'
character inside a frontend script would be displayed in non-interactive
mode as command entered by the user, leading to a confusing output.

Finally, some of the code to catch semicolon delimiters in commonds is
factored in between interactive and non-interactive modes.