[HandshakeToHW] Use `PortNameGenerator` everywhere
This makes all conversion patterns in the pass use the
`PortNameGenenrator` helper instead of manually querying the names of
all input and output ports. Previously, only the
`ExtModuleConversionnPattern` was using it.

At the same time, the commit improves the implementation of some parts
of the `PortNameGenerator` and adds port-name generation logic for all
supported arithmetic operations, yielding more meaningful names that
generic `in*` and `out*` for each.