[HandshakeToHW] Rework of external module creation
This contains an initial rework of the logic within the `HandshakeToHW`
conversion pass that creates external modules based on the specific
circuit needs. The code to identify "discriminating parameters" (name
pending, will change) is cleaned up and formalized, with the
introduction of data-structures (e.g., `ModuleDiscriminator`) to help in
that process. These data-structures will most likely be extracted from
the pass and moved to a generic support file in the future.

We attempt to start formalizing the interface between the conversion
pass and the future RTL emission backend, notably by introducing
operation attributes encoding the type of the original operation for
which the external module was created (`hw.name`) as well as the list of
parameters to concretize a specific implementation of that module. These
will most likely change in the future as we progress in the design of
our "RTL catalogue".

Additionally, we introduce new builders in the conversion pass to help
in the joint creation of external modules and instances for Handshake
operations. This improves code clarity and removes logic duplication
overall.

Finally, this commit also contains light changes to Handshake operations
and interfaces, needed to formalize the discriminating parameters for
certain operation types.