[docs] First specification for RTL generation
This adds a document which begins to describe our RTL generation
methodology, in particular the first formal iteration of a JSON format
to describe how to concretize real RTL components from their MLIR
equivalent, enabling the future backend to create full RTL designs for
our kernels using a mixture of generic RTL components and generated
ones. The document is incomplete but still meaningful as-is.