[MarkMemoryDependencies] Fix incorrect failure on op outside loops
This fixes a bug in the memory dependence analysis pass that would
fail incorrectly when trying to mark an out-of-loop write access
self-dependency. The method responsible for tagging the self-dependency
would (1) return failure instead of success on such occasions and (2)
fail to actually identify all of the operation's parent loops (it would
only check for affine loops but not scf loops).