[RTL] Add component/model matching logic
This adds methods to the `RTLConfiguration` and `RTLComponent` classes
to allow users to query for an RTL component or RTL model matching an
`hw::HWModuleExternOp` operation. The `RTLMatch` intermediate
data-strucuture defines the intended match's properties, and can be
automatically converted to from external module operations (making it
almost completely transparent for users).