[RTL] Refactor RTL "type system"
The in-memory representation of the lightweight type system used by
the RTL configuration was a bit verbose and hard to work with, and did
not enable clean verification of each parameter value's string encoding
as a separate step to constraint checking. This almost completely
changes the system to make it less verbose and more extensible, yielding
cleaner code and enabling external parts of the codebase  to access
encoding/decoding logic for each type statically.