[HandshakeToHW] Use new RTL parameter convention
Update to the conversion pass to make it use the newly established
convention (expected by RTL support) to pass RTL parameters from the
pass to the RTL backend. Instead of storing unnamed parameters in an
`ArrayAttr` on each external module, we use a dictionnary associating
each parameter name to its value. This makes parameter order irrelevant
and will allow more flexibility in the pattern matching system in the
future.