[RTL] Add "generic" key in parameter description
RTL parameters in the RTL configuration file now recognize the optional
"generic" key (mapped to a boolean value) to indicate whether they
represent a generic RTL parameter even when the associated component
is generated. This will let the RTL backend know that the parameter
value still needs to be supplied to the generated component's
instantiations. By default, the value is `false`.