[Backend] New text-based RTL generator
This introduces a generic generator for RTL components, for use in the
soon-to-come new RTL backend. It simply performs a user-provided number
of text replacements (with regex support) in an input RTL file, and
dumps the modified file's content to a specified location on disk. The
function performing text replacement will be used elsewhere in the
backend and is therefore added to the general RTL support compile unit.

The generator can be used for our "handshake.constant" RTL component,
whose RTL source is updated to work with the new generator.