[Backend] New control/store-less memory controller
This adds a version of the VHDL entity representing the memory
controller without any store or control-related ports. We need this
version to model Handshake memory controller operations that also have
no such ports; using the existing VHDL implementation would have led
to unbound IOs when instantiating the VHDL component from the new
backend (bi-dimensional vectors for control and store IOs would need to
have at least size 1, however we would not have any valid signal to
assign them to). The interface between the end synchronizer and the
memory controller---through the `memDone`--- signal, is also modified
to be a proper dataflow channel (though only the `valid` signal is
acttually used).