[LSQ] Extract gen. info derivation from DOT printer
The DOT printer extracts meta-information about LSQ's properties and
connectivity to the rest of the circuit to generate DOT attributes to
attach to each LSQ node. This information is then transformed into a
configuration file for the Chisel LSQ generator by the legacy VHDL
backend. The same extraction logic is needed by the new backend,
therefore this commit moves it to the generic helpers for memory
interfaces.