[Backend] Support for LSQ config. through JSON
The LSQ generator requires far more many parameters than all other
dataflow components, and expects them to be all encoded in a JSON file
whose path is provided in its arguments (as opposed to string-encoding
every individual parameter on the command line). This commit add
supports for this mechanism in our support library as well as in the
`HandshakeToHW` conversion pass. Key features are:
1. The ability to encode a restricted subset of MLIR attributes to JSON
   (in `Attributes.h/cpp`).
2. The ability for RTL configuration files to indicate that all RTL
   parameters should be serialized to JSON using the previously
   mentionned utility instead of being associated individually to
   parameter names (in `RTL.h/cpp`).
3. Support in the `HandshakeToHW` conversion pass to serialize all LSQ
   generation parameters to MLIR attributes, so they can be later
   serialized to JSON by the RTL backend for use by the LSQ generator
   (in `HandshakeToHW.cpp`).