[README] Add "setup" and "using" sections
Adds a dedicated section containing instructions on how to set up
Dynamatic (either from source or with the VM, whose setup instructions
are linked) and another short one linking to a couple
tutorials/documents important for first-time user.

This also adds the list of package dependencies for pacman-based Linux
distributions.