[RTL] Remove explicit copy-assignment constructor deletion
On older versions of clang (clang-14), this triggers a compilation error
because the return type `RTLType` is an abstract class, which makes
sense (though it works on more recent versions of clang mysteriously).
In any case, it is not necessary to explicitly delete the
copy-assignment constructor because the type is abstract in any case.