[docs] Add discussion on new type system proposal
Elaborates a bit more the discussion related to the proposed changes to
the type system to support extra channel signals or raw signals. This
includes an example on how core dataflow components do not structurally
change, a short note on how this will influence the RTL backend, and a
proposal for compose/decompose Handshake operations which could help
when lowering below Handshake.