[export-vhdl] Full remake of experimental RTL backend
This commit completely remakes the experimental RTL backend (that only
exports VHDL at this point, though it will not require much work to add
support for Verilog in the future) using the new RTL configuration file
format and code support. A new JSON-formatted configuration file
containing descriptions for all our currently supported RTL component
replaces the old RTL library, which is deleted. The backend is
structurally very simple, relying heavily on the existing RTL support
for the MLIR-to-RTL mapping logic.

There are still a few quirks to figure out before we can use this new
backend as the default one, but these shouldn't take long to iron out.