[docs] Complete RTL configuration file specification
This updates the RTL configuration file specification from its initial
version (which served as design proposal) to include all currently
supported options and syntax. Indirectly, this also serve as
documentation for the backend's RTL emission logic, which is alterable
through options in the RTL configuration file.

The document will most likely evolve to be a specification for the
entire backend. It currently contains two parts that will be soon
populated with information concerning the overall backends' design and
its internal matching logic.