[docs] Design section in backend/RTL documentation
This adds a large design/rationale section at the top of the
documentation for the backend/RTL configuration which covers the
overall behavior of the `HandshakeToHW` conversion pass as well as the
RTL emitter. The content is more akin to high-level documentation rather
than spec since all that it describes is already implemented.