[HandshakeToHW] `HWBuilder` class hierarchy
This brings some light improvements to the abstractions we use to
jointly create hardware instances and matching external modules in the
`HandshakeToHW` conversion pass. We now have a hierarchy of class rooted
in a generic `HWBuilder` that can be specialized to create specific
kinds of instance/module pairs depending on the needs. This offloads
some complexity from the conversion patterns themselves and allows the
creation of instance/modules "out of nowhere" (that do not originate
from an operation conversion).